# Parametrize CachingOptimizer and MockOptimizer on model type