Initial benchmark:
```
v0.6
126.218 ns (7 allocations: 176 bytes)
452.726 ns (17 allocations: 416 bytes)
 23.927 ns (3 allocations: 96 bytes)
 40.416 ns (5 allocations: 144 bytes)
v0.7
105.262 ns (5 allocations: 128 bytes)
414.565 ns (12 allocations: 304 bytes)
  9.623 ns (1 allocation: 48 bytes)
 29.113 ns (3 allocations: 96 bytes)
```
After parametrization of the MockOptimizer:
```
v0.6
121.149 ns (7 allocations: 176 bytes)
357.014 ns (14 allocations: 352 bytes)
 24.722 ns (3 allocations: 96 bytes)
 24.712 ns (3 allocations: 96 bytes)
v0.7
104.708 ns (5 allocations: 128 bytes)
311.206 ns (9 allocations: 240 bytes)
  9.306 ns (1 allocation: 48 bytes)
  9.400 ns (1 allocation: 48 bytes)
```
After parametrization of CachingOptimizer
```
v0.6
 38.889 ns (4 allocations: 112 bytes)
252.048 ns (11 allocations: 288 bytes)
 24.412 ns (3 allocations: 96 bytes)
 24.982 ns (3 allocations: 96 bytes)
v0.7
 23.149 ns (2 allocations: 64 bytes)
207.150 ns (6 allocations: 176 bytes)
  9.761 ns (1 allocation: 48 bytes)
  9.591 ns (1 allocation: 48 bytes)
```

Related to #321